Adding option control variables
When we present the options to the users we need to know which ones are required and further which ones are exclusive (choose one) vs inclusive (choose as many as you like). This data has been added to the manifest file and can be parsed (maybe... there's bug in the Inno Setup APIs that I have to circumvent first to test this).