krci.iss - Provided our own TStringList finder function
The TStringList class provides a function called Find() that is supposed to find a given string within the list if such a string is actually in the list. The function refuses to work properly, leading me to believe there's a bug in Inno Setup's implementation of it. As a result, I've written my own replacement function that does the same thing.