krci.iss - Improved Option handling
Attempting to understand which options the user selected through the GUI and preparing to download only those options exposed some bugs and weaknesses in the original option page setup and the manifest parsing. All of that has been fixed, option page setup has been simplified, and the option that the user selected is now properly queued up for downloading during the download phase.