Logging and manifest updates
The logging functions have all been replaced with a conditional logging function to limit the amount of clutter in the logs.

Furthermore, the manifest has been changed to use a more permanent cloud storage based on OwnCloud. As a result, some of the code to parse the filenames has been modified.